Moved class to separate file, added tests
